requirements.txt
langchain>=0.1.20
langchain-core>=0.1.45
langchain-google-genai>=1.0.3
langgraph>=0.0.45
google-generativeai>=0.5.0
python-dotenv>=1.0.0

